Rework tests to work with IPandoSerializer
