+++ tests/PandoTests/Tests/Repositories/TestStateTrees/TestTree.cs
@@ -1,25 +1,27 @@
 using System;
+using Pando.Serialization;
+using Pando.Serialization.Collections;
+using Pando.Serialization.Generic;
+using Pando.Serialization.Primitives;
 
 namespace PandoTests.Tests.Repositories.TestStateTrees;
 
-public record TestTree(string Name, TestTree.A MyA, TestTree.B MyB)
+public record TestTree(string Name, TestTree.A MyA, TestTree.B MyB) : IGenericSerializable<TestTree, string, TestTree.A, TestTree.B>
 {
-	public readonly struct A
-	{
-		public readonly int Age;
+	public static TestTree Construct(string name, A myA, B myB) => new(name, myA, myB);
 
-		public A(int age) { Age = age; }
+	public record A(int Age) : IGenericSerializable<A, int> {
+		public static A Construct(int age) => new(age);
 	}
 
-	public readonly struct B
-	{
-		public readonly DateTime Time;
-		public readonly int Cents;
-
-		public B(DateTime time, int cents)
-		{
-			Time = time;
-			Cents = cents;
-		}
+	public record B(DateTime Time, int Cents) : IGenericSerializable<B, DateTime, int> {
+		public static B Construct(DateTime time, int cents) => new(time, cents);
 	}
+
+	public static IPandoSerializer<TestTree> GenericSerializer() =>
+		new GenericNodeSerializer<TestTree, string, A, B>(
+			StringSerializer.ASCII,
+			new GenericNodeSerializer<A, int>(Int32LittleEndianSerializer.Default),
+			new GenericNodeSerializer<B, DateTime, int>(DateTimeToBinarySerializer.Default, Int32LittleEndianSerializer.Default)
+		);
 }
