+++ tests/PandoTests/Tests/Serialization/Primitives/StringSerializerTestData.cs
@@ -1,0 +1,8 @@
+using System.Text;
+using Pando.Serialization.Collections;
+using Pando.Serialization.PrimitiveSerializers;
+using Xunit;
+
+namespace PandoTests.Tests.Serialization.PrimitiveSerializers;
+
+
